refactor: messages and minor changes
