+++ src/predictions/profiles_mlcorelib/py_native/profiles_tutorial/file_generator.py
@@ -3,19 +3,17 @@ import logging
 import yaml
 
 from .config import TABLE_SUFFIX, CONFIG_FILE_PATH, INPUTS_FILE_PATH, PROFILES_FILE_PATH
-from .io_handler import IOHandler
 
 logger = logging.getLogger(__name__)
 
 
-class FileGenerator:
-    def __init__(self, fast_mode: bool, io_handler: IOHandler):
+class YamlGenerator:
+    def __init__(self, fast_mode: bool):
         self.fast_mode = fast_mode
         self.yaml = YAML()
         self.yaml.preserve_quotes = True
         self.yaml.indent(mapping=2, sequence=4, offset=2)
         self.yaml.width = 4096  # Prevent line wrapping
-        self.io_handler = io_handler
 
     def create_pb_project(self, entity_name, id_types, connection_name, id_graph_model):
         pb_project = {
